+++ changelog.txt
@@ -7,6 +7,7 @@ Added:
 - Added ability for some base Create contraptions to load cannons*
  - Mechanical Piston, Gantry, and Rope Pulley
 - Added Cannon Loader breaking features similar to Mechanical Piston
+- Added block hardness compatibility with Create's Copycat blocks
 Fixes:
 - Fixed drop mortar not dropping stored item if disassembled or broken before fired
 - Fixed drop mortar holding entire stack
