+++ common/src/main/java/rbasamoyai/createbigcannons/mixin/compat/create/FilterItemMixin.java
@@ -1,4 +1,4 @@
-package rbasamoyai.createbigcannons.mixin;
+package rbasamoyai.createbigcannons.mixin.compat.create;
 
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
